Add files via upload
Testing the usage of a shunt resistor to measure the relative current draw by the servo motors as a means to determine the minimum and maximum positions of the servo motors. The servo motors will draw a measurably greater amount of current when they encounter physical resistance to their movement (reaching the limits of the pan/tilt mechanism) and this measurement is used to determine the minimum and maximum positions to prevent excessive current draw and/or physical damage to the pan/tilt mechanism.